Create common module
