+++ component/common/src/main/java/com/example/otchallenge/common/ui/BaseDiffCallback.kt
@@ -1,0 +1,18 @@
+package com.example.otchallenge.common.ui
+
+import android.annotation.SuppressLint
+import androidx.recyclerview.widget.DiffUtil
+
+class BaseDiffCallback<T : Any>(
+    private val idExtractor: T.() -> Any?
+) : DiffUtil.ItemCallback<T>() {
+
+    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem.idExtractor() == newItem.idExtractor()
+    }
+
+    @SuppressLint("DiffUtilEquals")
+    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem == newItem
+    }
+}
