+++ feature/books-list/build.gradle
@@ -15,19 +15,16 @@ android {
 
 dependencies {
     implementation project(":component:base")
+    implementation project(":component:common")
 
     implementation libs.androidx.core.ktx
     implementation libs.androidx.appcompat
     implementation libs.material
-    implementation libs.androidx.activity
     implementation libs.androidx.constraintlayout
     implementation libs.androidx.swiperefreshlayout
     // dagger
     implementation libs.dagger
     kapt libs.dagger.compiler
-
-    //glide
-    implementation libs.glide
 
     // TODO migrate project to KSP
     //noinspection KaptUsageInsteadOfKsp
@@ -35,7 +32,6 @@ dependencies {
     implementation libs.moshi.moshi
 
     implementation libs.retrofit
-    implementation libs.timber
 
     testImplementation libs.junit
     testImplementation libs.kluent
