+++ gradle/libs.versions.toml
@@ -1,6 +1,7 @@
 [versions]
 agp = "8.4.2"
 kotlin = "1.9.21"
+kotlinxCoroutines = "1.9.0"
 coreKtx = "1.13.1"
 junit = "4.13.2"
 junitVersion = "1.1.5"
@@ -12,13 +13,13 @@ constraintlayout = "2.1.4"
 dagger = "2.51.1"
 retrofit = "2.11.0"
 glide = "4.16.0"
-rx-kotlin = "2.4.0"
 androidxNavigation = "2.8.4"
 swiperefreshlayout = "1.1.0"
 moshi = "1.15.1"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-recyclerview = { module="androidx.recyclerview:recyclerview", version = "1.3.2" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
@@ -38,7 +39,8 @@ glide = { group = "com.github.bumptech.glide", name = "glide", version.ref = "gl
 androidx-swiperefreshlayout = { group = "androidx.swiperefreshlayout", name = "swiperefreshlayout", version.ref = "swiperefreshlayout" }
 timber = { module = "com.jakewharton.timber:timber", version = "5.0.1" }
 kluent = { module = "org.amshove.kluent:kluent", version = "1.73"}
-kotlinx-coroutines-test = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-test", version="1.9.0" }
+kotlinx-coroutines-core = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref="kotlinxCoroutines" }
+kotlinx-coroutines-test = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-test", version.ref="kotlinxCoroutines" }
 mockito-kotlin = { module = "org.mockito.kotlin:mockito-kotlin", version = "5.4.0" }
 
 [plugins]
